Split out the trainer and generate all entities
