+++ run/meshMotion/wingMotion/wingMotion2D_pimpleFoam/system/controlDict
@@ -22,7 +22,7 @@ startTime       0;
 
 stopAt          endTime;
 
-endTime         0.1;
+endTime         0.2;
 //endTime         1e-2; // For debugging
 
 deltaT          1e-5;
